+++ pre_award/fund_store/api/routes.py
@@ -5,6 +5,7 @@ from distutils.util import strtobool
 from flask import abort, current_app, jsonify, request
 from fsd_utils.locale_selector.get_lang import get_lang
 
+from data.models import Round
 from pre_award.common.blueprints import Blueprint
 from pre_award.db import db
 from pre_award.fund_store.db.models.event import EventType
@@ -27,7 +28,6 @@ from pre_award.fund_store.db.schemas.event import EventSchema
 from pre_award.fund_store.db.schemas.fund import FundSchema
 from pre_award.fund_store.db.schemas.round import RoundSchema
 from pre_award.fund_store.db.schemas.section import SECTION_SCHEMA_MAP
-from services.data.models import Round
 
 fund_store_bp = Blueprint("fund_store_bp", __name__)
