+++ pre_award/fund_store/db/models/__init__.py
@@ -17,5 +17,4 @@ __all__ = [
     "SectionField",
     "FormName",
     "Event",
-    "ltree_extension",
 ]
