+++ pre_award/fund_store/scripts/data_updates/FS-2433-application_guidance.py
@@ -3,8 +3,8 @@ from sqlalchemy import update
 
 import pre_award.fund_store.config.fund_loader_config.cof.cof_r2 as cof_r2
 import pre_award.fund_store.config.fund_loader_config.cof.cof_r3 as cof_r3
+from data.models import Round
 from pre_award.db import db
-from services.data.models.round import Round
 
 
 def update_rounds_with_application_guidance(rounds):
