+++ pre_award/fund_store/scripts/data_updates/FS-3471-application_guidance.py
@@ -2,8 +2,8 @@ from flask import current_app
 from sqlalchemy import update
 
 import pre_award.fund_store.config.fund_loader_config.cyp.cyp_r1 as cyp_r1
+from data.models import Round
 from pre_award.db import db
-from services.data.models.round import Round
 
 
 def update_rounds_with_application_guidance(rounds):
