+++ pre_award/fund_store/scripts/data_updates/FS-3808_dpif_application_fields_download_available.py
@@ -2,8 +2,8 @@ from flask import current_app
 from sqlalchemy import update
 
 import pre_award.fund_store.config.fund_loader_config.digital_planning.dpi_r2 as dpi_r2
+from data.models import Round
 from pre_award.db import db
-from services.data.models.round import Round
 
 
 def update_application_fields_download_available(rounds):
