+++ pre_award/fund_store/scripts/data_updates/FS-4264_update_eoi_welsh_instructions.py
@@ -2,8 +2,8 @@ from flask import current_app
 from sqlalchemy import update
 
 import pre_award.fund_store.config.fund_loader_config.cof.eoi as eoi
+from data.models import Round
 from pre_award.db import db
-from services.data.models.round import Round
 
 
 def update_rounds_with_links(round_config):
