+++ tests/pre_award/fund_store_tests/conftest.py
@@ -13,11 +13,10 @@ from sqlalchemy_utils import Ltree
 from werkzeug.test import TestResponse
 
 from app import create_app
+from data.models import Fund, FundingType, Round
 from pre_award.config import Config
 from pre_award.fund_store.db.models.section import Section
 from pre_award.fund_store.db.queries import insert_fund_data, insert_sections, upsert_round_data
-from services.data.models.fund import Fund, FundingType
-from services.data.models.round import Round
 
 
 @pytest.fixture(scope="session")
