+++ tests/pre_award/fund_store_tests/test_data_fund_round.py
@@ -1,5 +1,6 @@
 import pytest
 
+from data.models import FundingType
 from pre_award.fund_store.db.queries import (
     get_all_funds,
     get_fund_by_id,
@@ -9,7 +10,6 @@ from pre_award.fund_store.db.queries import (
     get_rounds_for_fund_by_id,
     get_rounds_for_fund_by_short_name,
 )
-from services.data.models.fund import FundingType
 
 
 def test_get_fund_by_id(seed_dynamic_data):
